requirements.txt
# Core ML
tensorflow>=2.15.0
numpy>=1.24.0
scikit-learn>=1.3.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Frontend
streamlit>=1.31.0
pillow>=10.2.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Testing
pytest>=7.4.0
httpx>=0.26.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
